+++ src/main/java/com/dit/ebay/csv_model/CSVItem.java
@@ -18,10 +18,10 @@ public class CSVItem {
     private String description;
 
     @CsvBindByName(column = "buy_price")
-    private double buyPrice;
+    private BigDecimal buyPrice;
 
     @CsvBindByName(column = "first_bid")
-    private double firstBid;
+    private BigDecimal firstBid;
 
     @CsvBindByName(column = "time_ends")
     private Timestamp timeEnds;
@@ -71,19 +71,19 @@ public class CSVItem {
         this.description = description;
     }
 
-    public double getBuyPrice() {
+    public BigDecimal getBuyPrice() {
         return buyPrice;
     }
 
-    public void setBuyPrice(double buyPrice) {
+    public void setBuyPrice(BigDecimal buyPrice) {
         this.buyPrice = buyPrice;
     }
 
-    public double getFirstBid() {
+    public BigDecimal getFirstBid() {
         return firstBid;
     }
 
-    public void setFirstBid(double firstBid) {
+    public void setFirstBid(BigDecimal firstBid) {
         this.firstBid = firstBid;
     }
